Don't install libproj explicitly.
libproj-dev depends on its respective libproj library package,
no need to install it explicitly.

libproj-dev is a dependency of libgdal-dev, installing that is
sufficient.

The proj library package name has changed to reflect the SONAME bump
in PROJ 4.9.1-rc3. And will change again for PROJ 4.9.3.

libproj0 was used until PROJ 4.9.1-rc3.
libproj9 is currently used, and will change to libproj11 in 4.9.3.